Just another refactor
* Add abstract method to translate a user account viewmodel to it's
   appropriate DTO
* Update namespace
