+++ app/wama.core/ViewModel/User/EmployeeUserAccountViewModel.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using WAMA.Core.Models.DTOs;
+
+namespace WAMA.Core.ViewModel.User
+{
+    public class EmployeeUserAccountViewModel : UserAccountViewModel
+    {
+        public new UserAccountType AccountType { get; } = UserAccountType.Employee;
+
+        public override UserAccount ToUserAccount()
+        {
+            throw new NotImplementedException();
+        }
+    }
+}
