+++ app/wama.core/ViewModel/User/ManagerUserAccountViewModel.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using WAMA.Core.Models.DTOs;
+
+namespace WAMA.Core.ViewModel.User
+{
+    public class ManagerUserAccountViewModel : UserAccountViewModel
+    {
+        public new UserAccountType AccountType { get; } = UserAccountType.Manager;
+
+        public override UserAccount ToUserAccount()
+        {
+            throw new NotImplementedException();
+        }
+    }
+}
