+++ app/wama.core/ViewModel/User/MantainanceUserAccountViewModel.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using WAMA.Core.Models.DTOs;
+
+namespace WAMA.Core.ViewModel.User
+{
+    public class MantainanceUserAccountViewModel : UserAccountViewModel
+    {
+        public new UserAccountType AccountType { get; } = UserAccountType.Mantainance;
+
+        public override UserAccount ToUserAccount()
+        {
+            throw new NotImplementedException();
+        }
+    }
+}
