+++ app/wama.core/ViewModel/User/PatronUserAccountViewModel.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using WAMA.Core.Models.DTOs;
+
+namespace WAMA.Core.ViewModel.User
+{
+    public class PatronUserAccountViewModel : UserAccountViewModel
+    {
+        public new UserAccountType AccountType { get; } = UserAccountType.Patron;
+
+        public override UserAccount ToUserAccount()
+        {
+            throw new NotImplementedException();
+        }
+    }
+}
