--- app/wama.core/ViewModel/UserAccount/AdministratorUserAccountViewModel.cs
@@ -1,9 +1,0 @@
-﻿using WAMA.Core.Models.DTOs;
-
-namespace WAMA.Core.ViewModel.UserAccount
-{
-    public class AdministratorUserAccountViewModel : UserAccountViewModel
-    {
-        public new UserAccountType AccountType { get; } = UserAccountType.Administrator;
-    }
-}
