--- app/wama.core/ViewModel/UserAccount/EmployeeUserAccountViewModel.cs
@@ -1,9 +1,0 @@
-﻿using WAMA.Core.Models.DTOs;
-
-namespace WAMA.Core.ViewModel.UserAccount
-{
-    public class EmployeeUserAccountViewModel : UserAccountViewModel
-    {
-        public new UserAccountType AccountType { get; } = UserAccountType.Employee;
-    }
-}
