--- app/wama.core/ViewModel/UserAccount/ManagerUserAccountViewModel.cs
@@ -1,9 +1,0 @@
-﻿using WAMA.Core.Models.DTOs;
-
-namespace WAMA.Core.ViewModel.UserAccount
-{
-    public class ManagerUserAccountViewModel : UserAccountViewModel
-    {
-        public new UserAccountType AccountType { get; } = UserAccountType.Manager;
-    }
-}
