--- app/wama.core/ViewModel/UserAccount/MantainanceUserAccountViewModel.cs
@@ -1,9 +1,0 @@
-﻿using WAMA.Core.Models.DTOs;
-
-namespace WAMA.Core.ViewModel.UserAccount
-{
-    public class MantainanceUserAccountViewModel : UserAccountViewModel
-    {
-        public new UserAccountType AccountType { get; } = UserAccountType.Mantainance;
-    }
-}
