--- app/wama.core/ViewModel/UserAccount/PatronUserAccountViewModel.cs
@@ -1,9 +1,0 @@
-﻿using WAMA.Core.Models.DTOs;
-
-namespace WAMA.Core.ViewModel.UserAccount
-{
-    public class PatronUserAccountViewModel : UserAccountViewModel
-    {
-        public new UserAccountType AccountType { get; } = UserAccountType.Patron;
-    }
-}
